relay: refactoring
