+++ azurerm/internal/services/relay/validate/namespace_id.go
@@ -1,0 +1,22 @@
+package validate
+
+import (
+	"fmt"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/relay/parse"
+)
+
+// ValidateNamespaceID validates that the specified ID is a valid Relay Namespace ID
+func ValidateNamespaceID(i interface{}, k string) (warnings []string, errors []error) {
+	v, ok := i.(string)
+	if !ok {
+		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
+	}
+
+	if _, err := parse.NamespaceID(v); err != nil {
+		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
+		return
+	}
+
+	return warnings, errors
+}
